add prob20.py
I happened to see this repo of yours and I thought I'd play along. I saw that prob20 was missing, and I wasn't sure if you dismissed it as trivially easy or did the same as me, which was to fail to see the solution for a few minutes and, since it was tagged as easy, figured there must be an error in the description. But then I gave it another think and came up with this solution. Submitting for completeness sake.